agentic_insurance/requirements.txt
openai>=1.30.0
langgraph>=0.2.0
langchain-openai>=0.1.0
pydantic>=2.0.0
langfuse>=2.0.0
pytest>=8.0.0
python-dotenv>=1.0.0

